requirements.txt
Flask==2.1.2
azure-identity==1.6.0
azure-keyvault-secrets==4.4.0
